feat: refactor of utils
